Plot 2D Lattice
A lattice represents a crucial concept in mathematics and cryptographic applications, characterized by a periodic arrangement of points forming a grid.
Within the domain of cryptography, certain lattice problems serve as foundational elements for the development of highly secure cryptographic schemes. A notable example of such a problem is the Learning with Errors (LWE) problem and Ring Learning with Errors (RWLE).